backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0

# ML + RuBERT
transformers==4.39.3
tokenizers==0.15.2
torch==2.1.2

# YOLOv8
ultralytics==8.1.0

# Image processing
pillow==10.2.0
numpy==1.26.4

# Utils
pydantic==2.6.3
python-multipart==0.0.9

